Get GO gene sets

Get GO gene sets

getBioMartGOGeneSets(dataset, ontology = "BP",
    as_table = FALSE, gene_id_type = "ensembl_gene")

Arguments

dataset

A BioMart dataset. For a proper value, please see supportedOrganisms.

ontology

The value should be "BP", "CC", or "MF".

as_table

Whether to return the value as a data frame?

gene_id_type

Since BioMart is from Ensembl database, the default gene ID type is Ensembl gene ID. Depending on different organisms, Entrez ID ("entrez_gene") or gene symbol ("gene_symbol") can also be selected as the gene ID type.

Details

The gene sets are already compiled and are hosted on https://github.com/jokergoo/BioMartGOGeneSets_data , This function just simply retrieves data from there.

Value

A list of gene IDs or a data frame.

Examples

lt = getBioMartGOGeneSets("hsapiens_gene_ensembl")
lt = getBioMartGOGeneSets("hsapiens_gene_ensembl", gene_id_type = "entrez")
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
tb = getBioMartGOGeneSets("hsapiens_gene_ensembl", as_table = TRUE)